requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.1
websockets==11.0.3
wsproto==1.2.0
pandas==2.2.3
python-multipart==0.0.20
sqlalchemy==2.0.37
sqlalchemy-orm==1.2.10
psycopg2-binary==2.9.10
httpx==0.28.1
SQLAlchemy==2.0.37
requests==2.32.3
PyJWT==2.10.1
pillow==11.1.0
pyotp==2.9.0
aiohttp==3.11.12
lxml==5.3.0
fastapi-cache2==0.2.2
schedule==1.2.2
pygooglenews==0.1.3
passlib==1.7.4
python-jose==3.4.0
mibian==0.1.3
bcrypt==4.3.0
pyHanko==0.26.0
cachetools==5.5.2
openpyxl==3.1.5
apscheduler==3.11.0
email_validator==2.2.0
weasyprint==65.1
firebase-admin
aioboto3
reportlab
PyPDF2
